fix warnings

+++ vmt_driver/CommunicationManager.h
@@ -53,7 +53,7 @@ namespace VMTDriver {
 		static void OSCReceiver::SendAlive();
 		static void OSCReceiver::SendHaptic(int index, float frequency, float amplitude, float duration);
 		static void OSCReceiver::SendUnavailable(int code, std::string reason);
-		static void OSCReceiver::SendHmdInfo(float framerate, float screenWidth, float screenHeight, bool directMode);
+		static void OSCReceiver::SendHmdInfo(float framerate, int screenWidth, int screenHeight, bool directMode);
 		static void OSCReceiver::SendDevicePose(const Eigen::Affine3d& pose, const char* serialNumber);
 		static void OSCReceiver::SendDevices(std::string msg);
 	};
